# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SnowballR API",
        "description": "REST API for the SnowballR Application",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/users/me/reading-list/{paperId}": {
            "put": {
                "tags": [
                    "reading-list-controller"
                ],
                "operationId": "addPaperToReadingList",
                "parameters": [
                    {
                        "name": "paperId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "reading-list-controller"
                ],
                "operationId": "removePaperFromReadingList",
                "parameters": [
                    {
                        "name": "paperId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "head": {
                "tags": [
                    "reading-list-controller"
                ],
                "operationId": "isPaperOnReadingList",
                "parameters": [
                    {
                        "name": "paperId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
...
        "/api/v1/projects/{projectId}/invitations": {
            "get": {
                "tags": [
                    "invitation-controller"
                ],
                "operationId": "getPendingInvitationsForProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InvitationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "invitation-controller"
                ],
                "operationId": "inviteUserToProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InviteUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
...
        "/api/v1/invitations/{token}/accept": {
            "post": {
                "tags": [
                    "invitation-controller"
                ],
                "operationId": "acceptProjectInvitation",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InviteUserRequest": {
                "type": "object",
                "properties": {
                    "userEmail": {
                        "type": "string"
                    }
                }
            },
...
            "InvitationResponse": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DEFAULT",
                            "ADMIN"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE_UNCONFIRMED",
                            "ACTIVE",
                            "DELETED",
                            "CLEARED"
                        ]
                    }
                }
            },
...
        }
    }
}